Change function interface
* executeWithoutEither > execute
* createWithoutEither > create
* create > createWithEither
* execute > executeWithEither
* version bumps ctix@1.1.4
* apply dts-bundle-generator
* version bumps: 0.18.0 > 1.0.0
