+++ .circleci/config.yml
@@ -3,7 +3,7 @@ version: 2.1
 jobs:
   build:
     docker:
-      - image: circleci/node:12.16
+      - image: circleci/node:16.13.0
     steps:
       - checkout
       - restore_cache:
