+++ .ctiignore
@@ -1,6 +1,3 @@
 {
-  "**/__tests__": "*",
-  "**/__tests__/*": "*",
-  "**/__test__": "*",
-  "**/__test__/*": "*"
+  "**/__tests__/*": "*"
 }
