+++ jest.config.cjs
@@ -1,0 +1,9 @@
+module.exports = {
+  testEnvironment: 'node',
+  moduleFileExtensions: ['ts', 'tsx', 'js'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testMatch: ['**/__tests__/*.(ts|tsx)', '**/__test__/*.(ts|tsx)'],
+  setupFilesAfterEnv: ['./.config/jest.setup.cjs'],
+};
