+++ package.json
@@ -1,20 +1,21 @@
 {
   "name": "jin-frame",
-  "version": "0.18.0",
+  "version": "1.0.0",
   "description": "Repeatable HTTP request definition library",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "scripts": {
     "debug": "node --nolazy --require ts-node/register --inspect-brk=9229 ./node_modules/.bin/jest --no-cache --runInBand --detectOpenHandles",
-    "build": "just build",
-    "pub": "just pub",
-    "pub:prod": "just pub:prod",
-    "pub:dev": "just pub",
-    "clean": "just clean",
-    "lint": "just lint",
-    "test": "DEBUG=jinframe:* jest",
-    "just": "just",
-    "ctix": "ctix create -p ./tsconfig.json"
+    "build": "just --config ./.config/just.config.ts build",
+    "dts-bundle": "just --config ./.config/just.config.ts dts-bundle",
+    "pub": "just --config ./.config/just.config.ts pub",
+    "pub:prod": "just --config ./.config/just.config.ts pub:prod",
+    "pub:dev": "just --config ./.config/just.config.ts pub",
+    "clean": "just --config ./.config/just.config.ts clean",
+    "lint": "just --config ./.config/just.config.ts lint",
+    "test": "cross-env DEBUG=jinframe:* jest --config jest.config.cjs",
+    "just": "just --config ./.config/just.config.ts",
+    "ctix": "ctix create -p ./tsconfig.json --config ./.config/.ctirc"
   },
   "keywords": [
     "axios",
@@ -41,63 +42,51 @@
     "fp-ts": "^2.12.1",
     "http-status-codes": "^2.2.0",
     "luxon": "^2.4.0",
-    "my-easy-fp": "^0.12.0",
+    "my-easy-fp": "^0.13.0",
     "path-to-regexp": "^6.2.1",
     "reflect-metadata": "^0.1.13"
   },
   "devDependencies": {
-    "@tsconfig/node14": "^1.0.1",
+    "@tsconfig/node14": "^1.0.3",
     "@types/debug": "^4.1.7",
-    "@types/jest": "^28.1.1",
-    "@types/luxon": "^2.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.27.1",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.27.1",
-    "@typescript-eslint/parser": "^5.27.1",
+    "@types/jest": "^28.1.4",
+    "@types/luxon": "^2.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.30.4",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.30.4",
+    "@typescript-eslint/parser": "^5.30.4",
     "alias-help": "^0.1.0",
     "cross-env": "^7.0.3",
-    "ctix": "^1.0.3",
+    "ctix": "^1.1.4",
     "debug": "^4.3.4",
     "docsify-cli": "^4.4.4",
-    "eslint": "^8.17.0",
+    "dts-bundle-generator": "^6.12.0",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsdoc": "^39.3.2",
+    "eslint-plugin-jsdoc": "^39.3.3",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^28.1.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^28.1.2",
     "just-scripts": "^2.0.0",
-    "marked": "^4.0.16",
-    "nock": "^13.2.6",
-    "prettier": "^2.6.2",
+    "marked": "^4.0.17",
+    "nock": "^13.2.8",
+    "prettier": "^2.7.1",
     "prettier-eslint": "^15.0.1",
-    "prettier-plugin-organize-imports": "^2.3.4",
+    "prettier-plugin-organize-imports": "^3.0.0",
     "rimraf": "^3.0.2",
-    "ts-jest": "^28.0.4",
-    "ts-loader": "^9.3.0",
-    "ts-node": "^10.8.1",
+    "ts-jest": "^28.0.5",
+    "ts-loader": "^9.3.1",
+    "ts-node": "^10.8.2",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
     "tslint": "^6.1.3",
-    "typescript": "^4.7.3",
+    "typescript": "^4.7.4",
     "typescript-json-schema": "^0.53.1",
     "webpack": "^5.73.0",
-    "webpack-cli": "^4.9.2",
+    "webpack-cli": "^4.10.0",
+    "webpack-merge": "^5.8.0",
     "webpack-node-externals": "^3.0.0"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx)$": "ts-jest"
-    },
-    "testMatch": [
-      "**/__tests__/*.(ts|tsx)"
-    ]
   },
   "directories": {
     "test": "test"
